requirements.txt
python-dotenv>=1.0.0
openai>=1.0.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
pytest>=7.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
tiktoken>=0.5.1
chromadb>=0.4.22
python-jose>=3.3.0
tenacity>=8.2.3 